+++ Assets/HoloToolkit/Input/Scripts/Utilities/MicrophoneHelper.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 6c0764cefe2ea864081fdc8217d32a3c
-timeCreated: 1510613478
+guid: c9ac0991c94b1ab4881803bce4d496de
+timeCreated: 1510250600
 licenseType: Pro
 MonoImporter:
   externalObjects: {}
